pyproject.toml
[project]
name = "lucasys-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergey Andronnikov",email = "sergey.andronnikov@altexsoft.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright (>=0.7.0,<0.8.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "faker (>=37.4.0,<38.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
